Moved LocalizationSample to its own python module under theg.
