+++ theseus/utils/examples/__init__.py
@@ -3,7 +3,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .bundle_adjustment import ReprojectionError
+from .bundle_adjustment import LocalizationSample, ReprojectionError
 from .motion_planning import (
     InitialTrajectoryModel,
     MotionPlanner,
