+++ theseus/utils/examples/bundle_adjustment/__init__.py
@@ -1,1 +1,2 @@
+from .data import LocalizationSample
 from .reprojection_error import ReprojectionError
